requirements.txt
fastapi==0.135.3
langchain_core==1.2.28
langchain_groq==1.1.2
langfuse==4.2.0
langgraph==1.1.6
pydantic==2.12.5
python-dotenv==1.2.2
slowapi==0.1.9
supabase==2.28.3
tavily_python==0.7.23
langgraph-checkpoint-postgres==2.0.11
psycopg[binary]==3.2.9
langgraph-checkpoint-sqlite==2.0.6
uvicorn==0.34.0
python-multipart==0.0.20
langchain==0.3.25
langchain-community==0.3.24
streamlit
requests
